corrected pointer issues to build input tree
